Rename the top level folder from sailthru to edx_sailthru
