+++ edx_sailthru/sailthru_content/services/constants.py
