+++ edx_sailthru/sailthru_content/services/tests/test_catalog_api_service.py
@@ -1,7 +1,7 @@
 import responses
 
-from sailthru.sailthru_content.services.tests.catalog_api_test_mixins import CatalogApiTestMixins
-from sailthru.sailthru_content.services.tests.fixtures import (
+from catalog_api_test_mixins import CatalogApiTestMixins
+from fixtures import (
     SINGLE_COURSE_DATA, SINGLE_PROGRAM_DATA, SINGLE_SEARCHABLE_COURSE_DATA
 )
