+++ edx_sailthru/sailthru_content/services/tests/test_sailthru_api_service.py
@@ -3,7 +3,7 @@ import unittest
 
 from faker import Factory
 
-from sailthru.sailthru_content.services.sailthru_api_service import SailthruApiService
+from ..sailthru_api_service import SailthruApiService
 
 
 class SailthruApiServiceTests(unittest.TestCase):
